layout plan for simple fork setup
